% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{collapse_otu}
\alias{collapse_otu}
\title{Collapse otu table based on chosen level}
\usage{
collapse_otu(otu, metadata = mmn)
}
\arguments{
\item{otu}{otu table where species are rows and columns are stations}

\item{metadata}{metadata where all stations are rows}
}
\value{
collapsed otu table based on chosen level
}
\description{
This function for collapsing otu statoins (columns) based on n=column name in metadata
}
\examples{
otu<-collapse_otu #columns are stations and rows are species
}
