% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{colur}
\alias{colur}
\title{Generates colours that are not white}
\usage{
colur(n)
}
\arguments{
\item{n}{The number of colours to be produced}
}
\value{
R vector
}
\description{
This function randomly generates 'n' colours that are not white
}
\examples{
colur(4)
[1] "darkorange2" "orchid1"     "skyblue1"    "mediumblue"
}
