% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{rr}
\alias{rr}
\title{Remove blank rows from OTU and TAX}
\usage{
rr(df, c)
}
\arguments{
\item{df}{the input dataframe which the rows will be removed}

\item{c}{the logical vector expressing which rows should be removed}
}
\value{
Removes the the c rows
}
\description{
This function removes the blanks from OTU and TAX dataframe
}
\examples{
tax<-rr(tax,!c) #removing the c rows from tax data.frame
otu<-rr(otu,!c) #removing the c rows from otu data.frame
}
