% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.R
\name{write.result}
\alias{write.result}
\title{Prints out the statistical test in a .txt file}
\usage{
write.result(title, wd = getwd())
}
\arguments{
\item{title}{The name of the R object}

\item{wd}{wd=getwd(), if you want to change this you could write paste0(wd,"/",folder)}
}
\value{
Creates a .txt file in the wd(working directory) with the results
}
\description{
This function prints the statistical results in a text file with the name of the object as the name of the file.
}
\examples{
write.result(t.test(mydata))
write.result(adonis(dist ~ groups, permutations= 999))
}
